+++ jwst/tests_nightly/general/miri/test_fringe2.py
@@ -1,6 +1,5 @@
-import os
 import pytest
-from astropy.io import fits as pf
+from astropy.io import fits
 from jwst.fringe.fringe_step import FringeStep
 
 from ..helpers import add_suffix
@@ -20,19 +19,14 @@ def test_fringe_miri2(_bigdata):
     suffix = 'fringe'
     output_file_base, output_file = add_suffix('fringe2_output.fits', suffix)
 
-    try:
-        os.remove(output_file)
-    except:
-        pass
-
     FringeStep.call(_bigdata+'/miri/test_fringe/fringe2_input.fits',
                     output_file=output_file_base, suffix=suffix
                     )
-    h = pf.open(output_file)
-    href = pf.open(_bigdata+'/miri/test_fringe/baseline_fringe2.fits')
-    newh = pf.HDUList([h['primary'],h['sci'],h['err'],h['dq']])
-    newhref = pf.HDUList([href['primary'],href['sci'],href['err'],href['dq']])
-    result = pf.diff.FITSDiff(newh,
+    h = fits.open(output_file)
+    href = fits.open(_bigdata+'/miri/test_fringe/baseline_fringe2.fits')
+    newh = fits.HDUList([h['primary'],h['sci'],h['err'],h['dq']])
+    newhref = fits.HDUList([href['primary'],href['sci'],href['err'],href['dq']])
+    result = fits.diff.FITSDiff(newh,
                               newhref,
                               ignore_keywords = ['DATE','CAL_VER','CAL_VCS','CRDS_VER','CRDS_CTX'],
                               rtol = 0.00001