+++ jwst/tests_nightly/general/nirspec/test_spec2.py
@@ -6,7 +6,6 @@ import pytest
 
 from jwst.assign_wcs.util import NoDataOnDetectorError
 from jwst.pipeline import Spec2Pipeline
-from jwst.pipeline.collect_pipeline_cfgs import collect_pipeline_cfgs
 
 pytestmark = [
     pytest.mark.usefixtures('_jail'),
@@ -15,10 +14,6 @@ pytestmark = [
 ]
 
 
-@pytest.mark.xfail(
-    reason='Input data not available',
-    run=False
-)
 def test_nrs2_nodata_api(_bigdata):
     """
 
@@ -37,10 +32,6 @@ def test_nrs2_nodata_api(_bigdata):
         ))
 
 
-@pytest.mark.xfail(
-    reason='Input data not available',
-    run=False
-)
 def test_nrs2_nodata_strun(_bigdata):
     """Ensure that the appropriate exit status is returned from strun"""