+++ jwst/tests_nightly/general/pipelines/test_image2pipeline_2b.py
@@ -1,6 +1,5 @@
-import os
 import pytest
-from astropy.io import fits as pf
+from astropy.io import fits
 from jwst.pipeline.calwebb_image2 import Image2Pipeline
 
 pytestmark = [
@@ -18,11 +17,11 @@ def test_image2pipeline2b(_bigdata):
     Image2Pipeline.call(_bigdata+'/pipelines/jw82500001003_02101_00001_NRCALONG_rateints.fits',
                         output_file='jw82500001003_02101_00001_NRCALONG_calints.fits')
 
-    h = pf.open('jw82500001003_02101_00001_NRCALONG_calints.fits')
-    href = pf.open(_bigdata+'/pipelines/jw82500001003_02101_00001_NRCALONG_calints_ref.fits')
-    newh = pf.HDUList([h['primary'],h['sci'],h['err'],h['dq'],h['area']])
-    newhref = pf.HDUList([href['primary'],href['sci'],href['err'],href['dq'],href['area']])
-    result = pf.diff.FITSDiff(newh,
+    h = fits.open('jw82500001003_02101_00001_NRCALONG_calints.fits')
+    href = fits.open(_bigdata+'/pipelines/jw82500001003_02101_00001_NRCALONG_calints_ref.fits')
+    newh = fits.HDUList([h['primary'],h['sci'],h['err'],h['dq'],h['area']])
+    newhref = fits.HDUList([href['primary'],href['sci'],href['err'],href['dq'],href['area']])
+    result = fits.diff.FITSDiff(newh,
                               newhref,
                               ignore_keywords = ['DATE','CAL_VER','CAL_VCS','CRDS_VER','CRDS_CTX'],
                               rtol = 0.00001