+++ CHANGES.rst
@@ -82,6 +82,8 @@ fits_generator
 flatfield
 ---------
 
+- Modified the code to find the dispersion direction. [#2492]
+
 fringe
 ------